backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.4
psycopg[binary]==3.1.19
redis==5.0.6
celery==5.3.6
requests==2.32.3
openai==1.43.0
anthropic==0.36.1
pytest==8.3.2
PyJWT==2.9.0
